refactor: change float to bigdecimal

+++ src/main/java/libreria/modelo/bean/PedidoDetalle.java
@@ -1,5 +1,6 @@
 package libreria.modelo.bean;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class PedidoDetalle {
 
     private int cantidad;
-    private float subtotal;
+    private BigDecimal subtotal;
     private String codigo;
     private String SKU;
 }
